gymbuddy/requirements.txt
fastapi
uvicorn
anthropic
sqlite3
python-dotenv
jinja2
python-multipart
